refactor: enforce mypy --strict across all modules, split REPL and ChatView hotspots
Remove every project-level ignore_errors override: mypy --strict now
covers all 211 source files with zero exemptions. Root causes fixed
instead of ignores: QuoteCacheEntry/serializer typed as real Quote,
cache session() -> Iterator[Session], adapter Protocol signatures
aligned, web singleton getters typed (ConversationMemory/EpisodicMemory/
PredictionTracker/SemanticMemory/MemoryService), CLI factory helpers
return real types instead of object, _build_llm_client -> OpenAI | None.

Structural splits that came out of the convergence:
- cli.py 807 -> 518 lines: rendering, slash commands, agent Live view
  and workflow printing moved to cli_repl.py (strict typed); the REPL
  loop keeps only orchestration.
- tui/views/chat.py 946 -> 811 lines: 12 slash-command card builders
  extracted to SlashCardFactory in tui/views/slash_cards.py.

+++ src/mommy_chaogu/agent/service.py
@@ -465,7 +465,7 @@ class AgentService:
                         return max(0.0, float(retry_after))
                     except (TypeError, ValueError):
                         pass
-        return self._retry_base_delay * (2**attempt) + random.uniform(0, 0.5)
+        return float(self._retry_base_delay * (2**attempt) + random.uniform(0, 0.5))
 
     def _track_usage(self, response: Any, phase: str = "agent") -> None:
         """把一次 LLM 调用记进 TokenTracker（成本可观测性）。失败不阻塞主流程。"""
